+++ doc/changes.rst
@@ -5,7 +5,9 @@ Change Log
 1.0.6 (unreleased)
 ------------------
 
-* No changes yet.
+* Add additional functionality to Tucson and Utah mirror scripts (PR `#69`_).
+
+.. _`#69`: https://github.com/desihub/desitransfer/pull/69
 
 1.0.5 (2025-10-10)
 ------------------