+++ doc/changes.rst
@@ -5,6 +5,7 @@ Change Log
 1.0.6 (unreleased)
 ------------------
 
+* Ignore ``rrdetails-*.h5`` files at Utah.
 * Add additional functionality to Tucson and Utah mirror scripts (PR `#69`_).
 
 .. _`#69`: https://github.com/desihub/desitransfer/pull/69